#39 Implement filtering for list
