+++ demos/list/app/main.html
@@ -1,11 +1,13 @@
 <input id="searchbox" />
-<ig-list searchBoxId="searchbox">
+<ig-list >
+    <ig-header>Filtered List</ig-header>     
     <ig-item *ngFor="let navItem of navItems; let index = index" [options]="options">
       {{navItem.text}}
     </ig-item>
 </ig-list>
 
-<ig-list>
+<ig-list searchBoxId="searchbox">
+  <ig-header>Href List</ig-header>     
   <ig-item [href]="'http://google.com/'">Google</ig-item>
   <ig-item [href]="'http://youtube.com/'">YouTube</ig-item>
   <ig-item [href]="'http://facebook.com/'">Facebook</ig-item>
@@ -22,7 +24,7 @@
   <ig-item>Cosmic Crisp</ig-item>
   <ig-item>Pinova</ig-item>
 </ig-list>
-
+<h3>Non-header List</h3>
 <ig-list>
   <ig-item *ngFor="let navItem of navItems; let index = index">
     {{navItem.text}}
