+++ demos/list/app/main.js
@@ -26,8 +26,8 @@ AppComponent = __decorate([
         styleUrls: ["app/main.css"],
         templateUrl: "app/main.html",
         directives: [
-            Infragistics.Button,
-            Infragistics.Icon,
+            //Infragistics.Button,
+            //Infragistics.Icon,
             Infragistics.Header,
             Infragistics.Item,
             Infragistics.List
