+++ demos/tab/index.html
@@ -68,7 +68,7 @@
   <body>
     
     <div class="viewport">
-        <sample-app> Loading ... </sample-app>
+        <demo-app> Loading ... </demo-app>
     </div>
     <br>
     <div>
