+++ src/list/list-content.html
@@ -1,4 +1,4 @@
-<div #wrapper class="{{_innerStyle}}">
+<div #wrapper class="{{_innerStyle}}" [hidden]="hidden">
   <ng-content></ng-content>  
 </div>
 <div *ngIf="_panOptions" class="options">
