+++ src/main.ts
@@ -7,4 +7,5 @@ export * from './navigation-drawer/navigation-drawer'
 export * from './icon/icon'
 export * from './button/button'
 export * from './list/list'
+export * from './list/items'
 export * from './tab/tab'
